% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_projects.R
\name{condense_projects}
\alias{condense_projects}
\title{Condense projects}
\usage{
condense_projects(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Takes in flat table of metadata and extracts a subset of associated resource
information by project
}
