% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_abstract}
\alias{extract_abstract}
\title{Extract abstract column}
\usage{
extract_abstract(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{abstract} as a top level column.
}
\description{
Extract abstract column
}
