% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_citation}
\alias{extract_citation}
\title{Extract citation column}
\usage{
extract_citation(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{citation} and \code{title} as top level
columns.
}
\description{
Extract citation column
}
