% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_dataDictionary}
\alias{extract_dataDictionary}
\title{Extract dataDictionary column}
\usage{
extract_dataDictionary(mdFiles)
}
\arguments{
\item{mdFiles}{The \code{data} argument from the \code{flatten_mdJSON()}
function.}
}
\value{
The input tibble with \code{dataDictionary} as a top level column.
}
\description{
Extract dataDictionary column
}
