% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_dates}
\alias{extract_dates}
\title{Extract dates columns}
\usage{
extract_dates(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{startDateTime} and \code{endDateTime}
as top level columns. Creates columns \code{startDate} and \code{endDate}.
}
\description{
Extract dates columns
}
