% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_extent}
\alias{extract_extent}
\title{Extract extent column}
\usage{
extract_extent(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{extent} as a top level column.
}
\description{
Extract extent column
}
