% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_metadataInfo}
\alias{extract_metadataInfo}
\title{Extract metadataInfo column}
\usage{
extract_metadataInfo(mdFiles)
}
\arguments{
\item{mdFiles}{The \code{data} argument from the \code{flatten_mdJSON()}
function.}
}
\value{
The input tibble with \code{metadataInfo}, \code{metaId}, and
\code{namespace} as top level columns.
}
\description{
Extract metadataInfo column
}
