% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_people}
\alias{extract_people}
\title{Extract pointOfContact column}
\usage{
extract_people(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with all personnel roles as top level columns.
}
\description{
Extract pointOfContact column
}
