% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_resourceDistribution}
\alias{extract_resourceDistribution}
\title{Extract resourceDistribution column}
\usage{
extract_resourceDistribution(mdFiles)
}
\arguments{
\item{mdFiles}{The \code{data} argument from the \code{flatten_mdJSON()}
function.}
}
\value{
The input tibble with \code{resourceDistribtuion}, \code{contact},
\code{uri}, \code{onlineResourceName}, \code{function}, and
\code{transferSize} as top level columns.
}
\description{
Extract resourceDistribution column
}
