% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_resourceInfo}
\alias{extract_resourceInfo}
\title{Extract resourceInfo column}
\usage{
extract_resourceInfo(mdFiles)
}
\arguments{
\item{mdFiles}{The \code{data} argument from the \code{flatten_mdJSON()}
function.}
}
\value{
The input tibble with \code{resourceInfo} as a top level column.
}
\description{
Extract resourceInfo column
}
