% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_status}
\alias{extract_status}
\title{Extract status column}
\usage{
extract_status(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{status} as a top level column.
}
\description{
Extract status column
}
