% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_taxonomy}
\alias{extract_taxonomy}
\title{Extract taxonomy column}
\usage{
extract_taxonomy(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{taxonomicClassification} and
\code{taxonomicSystem} as top level columns.
}
\description{
Extract taxonomy column
}
