% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{extract_type}
\alias{extract_type}
\title{Extract resource type column}
\usage{
extract_type(resourceInfo)
}
\arguments{
\item{resourceInfo}{The output from the from the
\code{extract_resourceInfo()} function.}
}
\value{
The input tibble with \code{type} and \code{resourceName} as top
level columns.
}
\description{
Extract resource type column
}
