% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{flatten_mdJSON}
\alias{flatten_mdJSON}
\title{Flatten metadata}
\usage{
flatten_mdJSON(data)
}
\arguments{
\item{data}{A tibble of mdJSON files}
}
\value{
A partially flattened tibble of useful metadata fields for each
resource: title, responsible parties, resource type, dates, file location,
data dictionary, directory name, and associated resources.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Flattens a useful subset of metadata fields from imported mdJSON files to be
used in analysis and reports.
}
