% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Get Contacts}
\usage{
get_contacts(data)
}
\arguments{
\item{data}{Any number of mdJSON metadata records that have been flattened
into a tidy tibble}
}
\value{
A tibble containing all contacts found within the metadata records,
including their contactIds, names, and email addresses.
}
\description{
Get Contacts
}
