% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reformat_personnel.R
\name{reformat_personnel}
\alias{reformat_personnel}
\title{Reformat personnel}
\usage{
reformat_personnel(
  data,
  contacts,
  personnel = c("AKRegionDataTrustee", "AKRegionDataSteward", "AKRegionDataCustodian"),
  other_columns = c()
)
}
\arguments{
\item{other_columns}{}
}
\value{

}
\description{
Creates a list of essential data management personnel for each project with
their email address, along with any other requested columns.
}
