% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_project.R
\name{report_projects}
\alias{report_projects}
\title{Report project details}
\usage{
report_projects(data, contacts, project_list = data$projectName, output_dir)
}
\arguments{
\item{data}{Flattened metadata created by the function
\code{flatten_mdJSON()}.}

\item{contacts}{List of contacts from the metadata with their metadata IDs.}

\item{project_list}{List of projects that individual reports will be
generated for.}

\item{output_dir}{Directory that reports will be saved to.}
}
\value{
An html report containing project details.
}
\description{
Generates reports for individual projects.
}
