% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_mdJSON.R
\name{search_general}
\alias{search_general}
\title{General search}
\usage{
search_general(data, term)
}
\arguments{
\item{term}{}
}
\value{
A subset of resources that contain the specified term
}
\description{
Searches for records that contain a specified term in any part of the
metadata.
}
