% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_mdJSON.R
\name{search_taxon}
\alias{search_taxon}
\title{Search for a taxon}
\usage{
search_taxon(data, taxon)
}
\arguments{
\item{data}{Imported mdJSON data}

\item{taxon}{Taxon to search for}
}
\value{
A subset of resources that contain a specific taxon.
}
\description{
Searches for resources that contain a specific taxon.
}
