% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdJSON_extraction.R
\name{get_contacts}
\alias{get_contacts}
\title{Get metadata contacts}
\usage{
get_contacts(data)
}
\arguments{
\item{data}{A tibble of mdJSON files}
}
\value{
A tibble containing contact ID, contact name and email address for
contacts found in imported metadata.
}
\description{
Creates a table of contacts associated with metadata resources
}
