% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mdJSON.R
\name{import_mdJSON}
\alias{import_mdJSON}
\title{Import mdJSON metadata}
\usage{
import_mdJSON(
  program = "fes",
  path = "\\\\\\\\ifw7ro-file.fws.doi.net\\\\datamgt"
)
}
\arguments{
\item{program}{The program the report will be generated for. Choose from
"fes", "mbm", "nwrs", "osm", "sa", or "all".}

\item{path}{Path to directory that will be searched for metadata files.}
}
\value{
A tibble containing imported mdJSON metadata records. Each row is a
unique record
}
\description{
Gathers metadata files using the mdJSON metadata standard from a chosen
program directory (or all program directories) in the Alaska Regional Data
Repository into a tibble for further analysis in R. This function may take
several minutes to run, depending on the number of folders and JSON files in
the given directory.
}
