% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reformat_personnel.R
\name{reformat_personnel}
\alias{reformat_personnel}
\title{Reformat personnel}
\usage{
reformat_personnel(
  data,
  contacts,
  personnel = c("AKRegionDataTrustee", "AKRegionDataSteward", "AKRegionDataCustodian"),
  other_columns = c()
)
}
\arguments{
\item{data}{Tibble of flattened mdJSON resources}

\item{contacts}{Tibble of personnel extracted from the metadata}

\item{personnel}{List of personnel roles to be reformatted}

\item{other_columns}{Columns to be preserved, other than title, project name,
and the reformatted personnel roles.}
}
\value{
A tibble containing resource titles, project names, reformatted
personnel roles, and any other specified columns.
}
\description{
Creates a list of essential data management personnel for each project with
their email address, along with any other requested columns.
}
