% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reformat_personnel.R
\name{reformat_personnel}
\alias{reformat_personnel}
\title{Reformat personnel}
\usage{
reformat_personnel(data, contacts, personnel)
}
\arguments{
\item{data}{Tibble of flattened mdJSON resources}

\item{contacts}{Tibble of personnel extracted from the metadata}

\item{personnel}{List of personnel roles to be reformatted}
}
\value{
A tibble with reformatted personnel columns
}
\description{
For each data management role requested, changes the data from a nested list
to a comma separated list of personnel with their email addresses.
}
