% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_overview.R
\name{report_overview}
\alias{report_overview}
\title{Report overview of project metadata}
\usage{
report_overview(data, contacts, output_name, output_dir)
}
\arguments{
\item{data}{Subset of metadata created by \code{subset_mdJSON}.}

\item{contacts}{List of contacts from the metadata with their metadata IDs.}

\item{output_name}{Name to be given to the report generated.}

\item{output_dir}{Directory that the report should be generated in.}
}
\value{

}
\description{
Report overview of project metadata
}
