% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_overview.R
\name{report_overview}
\alias{report_overview}
\title{Report overview of project metadata}
\usage{
report_overview(
  data,
  contact_person = "principalInvestigator",
  essential_personnel = c("AKRegionDataTrustee", "AKRegionDataSteward",
    "AKRegionDataCustodian"),
  output_name,
  output_dir
)
}
\arguments{
\item{data}{Imported mdJSON files.}

\item{contact_person}{The role for the project's point of contact}

\item{essential_personnel}{Essential personnel titles for projects.}

\item{output_name}{Name to be given to the report generated.}

\item{output_dir}{Directory that the report should be generated in.}
}
\value{
An html report containing project details.
}
\description{
Generates an html report showing details on all projects for a single
program.
}
\details{
Reports contain a list of all projects, the number and types of data
products, a searchable table of all ongoing project details, and a table of
critical project personnel.
}
